Bump crypto package
There is a CVE in the crypto package used by the kubernetes libraries.
This is not fixed in the 0.17.x releases and the 0.18.x release
introduces breaking changes that affect kube-prompt. Therefore, the most
straightforward way to ensure that the crypto package is updated is to
use the replace directive.